# docs/swagger.yaml
basePath: /
definitions:
  api.loginUserRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  api.registerUserRequest:
    properties:
      is_admin:
        default: false
        type: boolean
      password:
        type: string
      username:
        type: string
    type: object
...
  postgres.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      updated_at:
        type: string
      username:
        type: string
    type: object
  token.Token:
    properties:
      expiry:
        type: string
      token:
        type: string
    type: object
  utils.Envelope:
    additionalProperties: {}
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a simple online judge
  title: Adel Online Judge API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Get list of all users (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                users:
                  items:
                    $ref: '#/definitions/postgres.User'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - admin
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - admin
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with username and password, returns JWT token
      parameters:
      - description: User login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/api.loginUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                token:
                  $ref: '#/definitions/token.Token'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
      summary: Login user
      tags:
      - auth
  /users/logout:
    delete:
      consumes:
      - application/json
      description: Logout user by deleting all their authentication tokens
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                message:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /users/me:
    get:
      consumes:
      - application/json
      description: Get the current authenticated user's information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                user:
                  $ref: '#/definitions/postgres.User'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Register a new normal user account
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/api.registerUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                user:
                  $ref: '#/definitions/postgres.User'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
      summary: Register user
      tags:
      - users
  /users/register-admin:
    post:
      consumes:
      - application/json
      description: Register a new admin user account
      parameters:
      - description: Admin user registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/api.registerUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                user:
                  $ref: '#/definitions/postgres.User'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Envelope'
            - properties:
                error:
                  type: string
              type: object
      summary: Register admin user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"